Extrais tests de DepotDonneesAutorisations
